pyproject.toml
# pyproject.toml is a new configuration file defined in PEP 518.
# It is design to store build system requirements, but it can also store any
# tool configuration for your Python project, possibly replacing the need for
# setup.cfg or other tool-specific files.
# https://github.com/carlosperate/awesome-pyproject
[build-system]
requires = ["maturin>=1.0,<2.0", "setuptools"]
build-backend = "maturin"

[project]
name = "sevirs"
version = "0.0.1"
requires-python = ">=3.10"
description = "Interface to SEVIR dataset"
authors = [{ name = "Jason Leaver", email = "jason.leaver@us.af.mil" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
]
dependencies = [
    "torch==2.0.1",
    "numpy==1.25.1",
    "polars==0.18.7",
    "pyarrow==12.0.1",
    "h5py==3.9.0",
    "tqdm==4.65.0",
    "typing_extensions==4.7.1",
    "scipy==1.11.1",
    # - extracting mrms and probsevere data
    "pandas==2.0.3",
    "requests==2.31.0",
    "lxml==4.9.3",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "ruff",
    "black",
    "isort",
    "tensor-annotations",
    "mypy-extensions",
    "pyarrow-stubs",
]
ipynb = ["ipykernel", "black", "jupyter-black"]
all = [
    # dev tools
    "mypy",
    "pytest",
    "ruff",
    "black",
    "isort",
    "types-tqdm",
    "mypy-extensions",
    "tensor-annotations-tensorflow-stubs",
    "pyarrow-stubs",
    "pandas-stubs",
    # "jupyter"
    "ipykernel",
    "black",
    "jupyter-black",
]

[tool.maturin] # https://www.maturin.rs/
# "extension-module" tells pyo3 we want to build an extension module (skips linking against libpython.so)
# features = ["pyo3/extension-module"]
features = ["pyo3/extension-module"]
module-name = "sevirs._lib"

[tool.mypy]
python_version = "3.10"
platform = "linux"
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
enable_incomplete_feature = "TypeVarTuple,Unpack"
exclude = ["docs/"]

[tool.pytest]
testpaths = ["tests/"]

[tool.coverage.run]
branch = true
omit = ["sevirs/_typing.py"]
source = ["sevirs/"]

[tool.coverage.report]
ignore_errors = false
show_missing = true
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Don't complain about missing debug-only code:s
    # "def __repr__",
    # "if self.debug",
    # Don't complain if tests don't hit defensive assertion code:
    # "raise AssertionError",
    "raise NotImplementedError",
    "AbstractMethodError",
    # Don't complain if non-runnable code isn't run:
    # "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
target-version = ['py310']
line-length = 119
include_trailing_comma = false

[tool.isort]
profile = "black"

[tool.ruff]
target-version = "py310"
line-length = 119
fix = true
ignore = [
    "E731", # do not assign a lambda expression, use a def
    "E402", # module level import not at top of file
    "E402", # module level import not at top of file
    "E501", # line too long
    "E741", # do not use variables named 'l', 'O', or 'I'
]
